+++ mesos/heatbeat.go
@@ -28,6 +28,7 @@ func Heartbeat() {
 		if task.State == "" {
 			framework.CommandChan <- task
 
+			mesosutil.Revive()
 			task.State = "__NEW"
 			data, _ := json.Marshal(task)
 			err := config.RedisClient.Set(config.RedisCTX, task.TaskName+":"+task.TaskID, data, 0).Err()
@@ -38,13 +39,14 @@ func Heartbeat() {
 		}
 
 		if task.State == "__NEW" {
-			mesosutil.Revive()
 			suppress = false
+			config.Suppress = false
 		}
 	}
 
-	if suppress {
+	if suppress && !config.Suppress {
 		mesosutil.SuppressFramework()
+		config.Suppress = true
 	}
 }